Create enventHandler.js
This JavaScript code adds functionality to the "START" button on a web page. When the button is clicked, the selected files are asynchronously uploaded to the server using the XMLHttpRequest object. A second request is then sent to process the downloaded files. After successful processing, the "START" button changes to a confirmation icon and the "Download" text element appears, notifying the user that the results are ready to be downloaded.