Create eventHandler_2.js
This JavaScript code adds functionality to the "START" button for the second type of downloadable file on a web page. When the button is clicked, the selected files are asynchronously uploaded to the server using the fetch object. A request is then sent to process the uploaded files. After successful processing, the "START" button changes to a confirmation icon, and the "Download" text element appears, notifying the user that the results are ready to download.